Refactoring for the map view
There were a number of "special cases" for the consensus weight vs.
advertised bandwidth that this commit tries to reduce. As a result the
semantics of the map analysis have changed although not with any
noticable difference to the results.

In the previous version the value for each country was determined by the
total of the consensus weight to the total of the advertised bandwidth.
This commit changes this to make the value for each country a mean
average of the ratios for individual relays, and now excludes relays
with an advertised bandwidth of 0 (or missing advertised bandwidth) and
those relays that are not yet measured by at least 3 bandwidth
authorities.

During refactoring, tooltips have also been added to display values when
hovering over a country.